refactor(backend): split publications/endpoints.py into sub-package
Decomposes the 690-LOC flat publications/endpoints.py into a
sub-package at the same import path:

  publications/endpoints/
    __init__.py         (29)   aggregator router mounted at /api/v2/publications
    schemas.py          (94)   Pydantic models (AuthorModel, PublicationMetadataResponse, PublicationListItem, SyncResponse)
    metadata_route.py   (149)  GET /{pmid}/metadata
    list_route.py       (298)  GET / (JSON:API paginated list with filters/sort/search)
    sync_route.py       (147)  POST /sync + _sync_publications_background

Every file is under 500 LOC. Public import path unchanged, so
app.main's ``from app.publications import endpoints as
publication_endpoints`` + ``app.include_router(publication_endpoints.router)``
continues to work without modification.

list_route factors out ``_build_where_clauses`` as a shared helper
between the list and count queries — the pre-Wave-4 monolith had
the same clauses duplicated twice, with a subtle risk that the two
would drift. The new helper guarantees they stay in sync.

HTTP surface byte-identical. Counts unchanged: 907 passed, 9
skipped, 3 xfailed.
